+++ lib/src/livecurrencyrate.dart
@@ -13,7 +13,11 @@ class LiveCurrencyRate {
       String currentCurrency, String toCurrency, double price) async {
     final uri = Uri.parse(
         'https://api.softasium.com/Currency/$currentCurrency/$toCurrency/$price');
-    final headers = {"Authorization": "iamsyedidrees", "Content-Lenght": "0"};
+    final headers = {
+      "Authorization": "iamsyedidrees",
+      "Content-Lenght": "0",
+      "Content-Type": "text/html",
+    };
     // Map<String, dynamic> body = {'id': 21, 'name': 'bob'};
     // String jsonBody = json.encode(body);